Corrigidos pequenos erros.

+++ src/br/inf/ufes/ppd/TrabUtils.java
@@ -204,7 +204,7 @@ public class TrabUtils {
 		return false;
 	}
         
-        public static void Resultados(String arq, int tam, long time) throws IOException {
+        public static void Resultados(String arq, int tam, double time) throws IOException {
 		BufferedWriter buffWrite = new BufferedWriter(new FileWriter(arq, true));
         buffWrite.append(tam + "," + time + System.getProperty("line.separator"));
         buffWrite.close();
